EBP-19: As a Solace Developer, I can use cache containers in integration testing.
Removed debug print statements that were necessary only for development.
Refactored location of SEMP calls etc. so that cache functionality is in
isolated module that can be accessed by multiple test contexts.
Confirmed that as of this commit, cache tests and configuration can be
run either against an existing broker, or against a new broker.